+++ lib/opticore.dart
@@ -1,4 +1,21 @@
 //Config
+export 'package:bot_toast/bot_toast.dart';
+export 'package:cached_network_image/cached_network_image.dart';
+export 'package:connectivity_plus/connectivity_plus.dart';
+export 'package:dio/dio.dart';
+export 'package:equatable/equatable.dart';
+// Dependencies
+export 'package:flutter_bloc/flutter_bloc.dart';
+export 'package:flutter_svg/flutter_svg.dart';
+export 'package:font_awesome_flutter/font_awesome_flutter.dart';
+export 'package:internet_connection_checker_plus/internet_connection_checker_plus.dart';
+export 'package:lottie/lottie.dart';
+export 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+export 'package:talker/talker.dart';
+export 'package:talker_dio_logger/talker_dio_logger.dart';
+export 'package:talker_dio_logger/talker_dio_logger_interceptor.dart';
+export 'package:talker_dio_logger/talker_dio_logger_settings.dart';
+
 export 'src/config/config_import.dart';
 // Core
 export 'src/core/base/import/base_import.dart';
@@ -13,21 +30,3 @@ export 'src/services/services_import.dart';
 export 'src/utils/logger/logger_import.dart';
 // Util
 export 'src/utils/util_import.dart';
-
-
-// Dependencies
-export 'package:flutter_bloc/flutter_bloc.dart';
-export 'package:talker/talker.dart';
-export 'package:talker_dio_logger/talker_dio_logger.dart';
-export 'package:talker_dio_logger/talker_dio_logger_interceptor.dart';
-export 'package:talker_dio_logger/talker_dio_logger_settings.dart';
-export 'package:dio/dio.dart';
-export 'package:bot_toast/bot_toast.dart';
-export 'package:font_awesome_flutter/font_awesome_flutter.dart';
-export 'package:lottie/lottie.dart';
-export 'package:equatable/equatable.dart';
-export 'package:flutter_svg/flutter_svg.dart';
-export 'package:internet_connection_checker_plus/internet_connection_checker_plus.dart';
-export 'package:connectivity_plus/connectivity_plus.dart';
-export 'package:cached_network_image/cached_network_image.dart';
-export 'package:modal_bottom_sheet/modal_bottom_sheet.dart';