+++ lib/src/config/infrastructure/core_setup.dart
@@ -1,15 +1,15 @@
 part of '../config_import.dart';
 
-/// `CoreSetup` is a widget responsible for initializing and configuring essential global settings for the application.
+/// [CoreSetup] is a widget responsible for initializing and configuring essential global settings for the application.
 /// It manages core aspects such as the theme, localization, routing, and screens for handling errors like Maintenance,
 /// No Internet, and Not Found pages. This widget is typically used as the root widget to ensure the application is set up
 /// with the correct configurations and runs smoothly with the desired settings.
 ///
-/// The `CoreSetup` widget facilitates customization for error screens, allowing developers to display custom content or
+/// The [CoreSetup] widget facilitates customization for error screens, allowing developers to display custom content or
 /// behaviors when the app encounters maintenance mode, no internet connection, or missing pages.
 ///
 /// **Constructor Parameters:**
-/// - `appConfig`: The configuration object that contains the global settings for the app. This includes theme data,
+/// - [appConfig]: The configuration object that contains the global settings for the app. This includes theme data,
 ///   localization delegates, initial routes, and more. This parameter is **required** and ensures the app is initialized
 ///   with consistent settings.
 ///   Example:
@@ -20,7 +20,7 @@ part of '../config_import.dart';
 ///     initialRoute: '/home',
 ///   )
 ///   ```
-/// - `maintenanceConfig` (optional): A custom configuration object for the maintenance screen. If provided, it overrides
+/// - [maintenanceConfig] (optional): A custom configuration object for the maintenance screen. If provided, it overrides
 ///   the default maintenance screen configuration. If not provided, the default behavior will be used.
 ///   Example:
 ///   ```dart
@@ -28,7 +28,7 @@ part of '../config_import.dart';
 ///     customMessage: 'We are under maintenance, please try again later.',
 ///   )
 ///   ```
-/// - `noInternetConfig` (optional): A custom configuration for the no internet screen. If provided, it customizes the
+/// - [noInternetConfig] (optional): A custom configuration for the no internet screen. If provided, it customizes the
 ///   appearance and behavior of the no internet page. If not provided, the app will default to its standard behavior.
 ///   Example:
 ///   ```dart
@@ -36,7 +36,7 @@ part of '../config_import.dart';
 ///     customAnim: 'assets/no_internet_anim.json',
 ///   )
 ///   ```
-/// - `notFoundConfig` (optional): A custom configuration for the not found screen that appears when a user navigates to
+/// - [notFoundConfig] (optional): A custom configuration for the not found screen that appears when a user navigates to
 ///   a non-existent route. If provided, it customizes the behavior of this screen. If not provided, the default screen
 ///   will be shown.
 ///   Example:
@@ -44,6 +44,14 @@ part of '../config_import.dart';
 ///   NotFoundConfig(
 ///     customAnim: 'assets/404_animation.json',
 ///   )
+///   ```
+///   - [onBeforeConfigApply] (optional): A callback function that runs before applying the configurations. This can be used
+///   to perform any necessary setup or initialization tasks before the app configurations are applied.
+///   Example:
+///   ```dart
+///   onBeforeConfigApply: () async {
+///   await Firebase.initializeApp();
+///   },
 ///   ```
 class CoreSetup extends StatefulWidget {
   /// The configuration object that holds the global settings for the application.
@@ -61,6 +69,9 @@ class CoreSetup extends StatefulWidget {
   /// Custom configuration for the Not Found screen. If not provided, the default behavior will be used.
   final NotFoundConfig? notFoundConfig;
 
+  /// A callback to handle Firebase or other initialization before configurations are applied.
+  final Future<void> Function()? onBeforeConfigApply;
+
   /// Constructor for initializing the `CoreSetup` widget with app configurations.
   const CoreSetup({
     super.key,
@@ -68,21 +79,37 @@ class CoreSetup extends StatefulWidget {
     this.maintenanceConfig,
     this.noInternetConfig,
     this.notFoundConfig,
+    this.onBeforeConfigApply,
   });
 
   @override
   State<CoreSetup> createState() => _CoreSetupState();
 }
 
-class _CoreSetupState extends State<CoreSetup> {
+class _CoreSetupState extends State<CoreSetup> with AfterLayoutMixin {
   /// Function to initialize BotToast for displaying toast notifications in the app.
   ///
   /// This builder wraps widgets in the app and allows toast notifications to appear.
   final Function(BuildContext, Widget?) defaultBotToastBuilder = BotToastInit();
 
+  /// Initializes the configurations for the app, including custom error screens.
+  ///
+  /// This method is called after the first layout to ensure the app is fully initialized.
   @override
-  void initState() {
-    super.initState();
+  FutureOr<void> afterFirstLayout(BuildContext context) {
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (mounted) {
+        _initializeConfigurations();
+      }
+    });
+  }
+
+  /// Initializes Firebase or other resources and then applies custom configurations.
+  Future<void> _initializeConfigurations() async {
+    if (widget.onBeforeConfigApply != null) {
+      await widget.onBeforeConfigApply!();
+    }
+
     _applyCustomConfigurations();
   }
 
@@ -109,13 +136,18 @@ class _CoreSetupState extends State<CoreSetup> {
       onTap: () {
         // Dismiss the keyboard when tapping outside the text fields.
         FocusScopeNode currentFocus = FocusScope.of(context);
-        if (!currentFocus.hasPrimaryFocus && currentFocus.focusedChild != null) {
+        if (!currentFocus.hasPrimaryFocus &&
+            currentFocus.focusedChild != null) {
           FocusManager.instance.primaryFocus!.unfocus();
         }
       },
       child: MaterialApp(
         theme: widget.appConfig.theme,
+        scrollBehavior:
+            widget.appConfig.scrollBehavior ?? ScrollBehaviorHelper(),
+        scaffoldMessengerKey: widget.appConfig.scaffoldMessengerKey,
         title: widget.appConfig.appTitle ?? 'OptiCore',
+        locale: widget.appConfig.locale,
         builder: (context, child) {
           // Apply the BotToast builder for toast notifications, if provided.
           child = defaultBotToastBuilder(context, child);
@@ -153,11 +185,15 @@ class _CoreSetupState extends State<CoreSetup> {
           BotToastNavigatorObserver(),
           LoggerRouterObserver(),
         ],
-        onUnknownRoute: widget.appConfig.onUnknownRoute ?? _defaultRouteGenerator,
-        showPerformanceOverlay: widget.appConfig.showPerformanceOverlay ?? false,
+        onUnknownRoute:
+            widget.appConfig.onUnknownRoute ?? _defaultRouteGenerator,
+        showPerformanceOverlay:
+            widget.appConfig.showPerformanceOverlay ?? false,
         showSemanticsDebugger: widget.appConfig.showSemanticsDebugger ?? false,
-        checkerboardRasterCacheImages: widget.appConfig.checkerboardRasterCacheImages ?? false,
-        checkerboardOffscreenLayers: widget.appConfig.checkerboardOffscreenLayers ?? false,
+        checkerboardRasterCacheImages:
+            widget.appConfig.checkerboardRasterCacheImages ?? false,
+        checkerboardOffscreenLayers:
+            widget.appConfig.checkerboardOffscreenLayers ?? false,
       ),
     );
   }