+++ lib/src/config/settings/app_config.dart
@@ -34,6 +34,8 @@ part of '../config_import.dart';
 /// - [checkerboardRasterCacheImages]: A flag to enable or disable the checkerboard raster cache images for
 ///   debugging. This feature helps to identify performance bottlenecks related to rasterization of images.
 /// - [checkerboardOffscreenLayers]: A flag to enable or disable the checkerboard offscreen layers for debugging.
+/// - [scrollBehavior]: The scroll behavior for the app, allowing customization of the scrolling behavior.
+/// - [scaffoldMessengerKey]: The key for the scaffold messenger, which is used to show snack bars and other messages.
 ///   This helps to detect issues with offscreen rendering layers, such as those involved in animations or scrolling.
 class AppConfig extends Equatable {
   /// The app theme, which is customizable.
@@ -87,6 +89,18 @@ class AppConfig extends Equatable {
   /// This is useful for detecting issues with offscreen rendering layers.
   final bool? checkerboardOffscreenLayers;
 
+  /// The scroll behavior for the app.
+  /// This allows customizing the scroll behavior of the app.
+  final MaterialScrollBehavior? scrollBehavior;
+
+  /// The key for the scaffold messenger.
+  /// This key can be used to show snack bars and other messages.
+  final GlobalKey<ScaffoldMessengerState>? scaffoldMessengerKey;
+
+  /// The current locale of the app.
+  /// This locale determines the language and region settings for the app.
+  final Locale? locale;
+
   /// Constructor for the [AppConfig] class. Initializes the configuration properties.
   const AppConfig({
     this.theme,
@@ -102,22 +116,28 @@ class AppConfig extends Equatable {
     this.showSemanticsDebugger,
     this.checkerboardRasterCacheImages,
     this.checkerboardOffscreenLayers,
+    this.scrollBehavior,
+    this.scaffoldMessengerKey,
+    this.locale,
   });
 
   @override
   List<Object?> get props => [
-    theme,
-    appTitle,
-    localizationsDelegates,
-    supportedLocales,
-    initialRoute,
-    onGenerateRoute,
-    navigatorObservers,
-    color,
-    onUnknownRoute,
-    showPerformanceOverlay,
-    showSemanticsDebugger,
-    checkerboardRasterCacheImages,
-    checkerboardOffscreenLayers,
-  ];
+        theme,
+        appTitle,
+        localizationsDelegates,
+        supportedLocales,
+        initialRoute,
+        onGenerateRoute,
+        navigatorObservers,
+        color,
+        onUnknownRoute,
+        showPerformanceOverlay,
+        showSemanticsDebugger,
+        checkerboardRasterCacheImages,
+        checkerboardOffscreenLayers,
+        scrollBehavior,
+        scaffoldMessengerKey,
+        locale,
+      ];
 }