+++ lib/src/core/base/import/base_import.dart
@@ -1,13 +1,10 @@
 import 'dart:io';
 
-import 'package:bot_toast/bot_toast.dart';
-import 'package:equatable/equatable.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:opticore/opticore.dart';
 import 'package:opticore/src/core/base/widget/appbar/core_app_bar.dart';
 import 'package:opticore/src/core/base/widget/exception_widget.dart';