+++ lib/src/extensions/extensions_import.dart
@@ -1,7 +1,5 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:opticore/opticore.dart';
 
 part '../extensions/data/context_extension.dart';
@@ -12,5 +10,5 @@ part '../extensions/data/map_extension.dart';
 part '../extensions/data/navigation_extension.dart';
 part '../extensions/data/padding_extension.dart';
 part '../extensions/data/precache_extension.dart';
-part '../extensions/data/string_extension.dart';
 part '../extensions/data/set_non_null_extension.dart';
+part '../extensions/data/string_extension.dart';