+++ lib/src/reusable/mixin/after_layout_mixin.dart
@@ -1,0 +1,15 @@
+part of '../reusable_import.dart';
+
+mixin AfterLayoutMixin<T extends StatefulWidget> on State<T> {
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.endOfFrame.then(
+      (_) {
+        if (mounted) afterFirstLayout(context);
+      },
+    );
+  }
+
+  FutureOr<void> afterFirstLayout(BuildContext context);
+}