+++ lib/src/utils/util_import.dart
@@ -1,15 +1,15 @@
-import 'package:bot_toast/bot_toast.dart';
-import 'package:connectivity_plus/connectivity_plus.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:internet_connection_checker_plus/internet_connection_checker_plus.dart';
 import 'package:opticore/opticore.dart';
 
 import 'ui/core_colors.dart';
 
 part '../utils/helper/connection_helper.dart';
+part '../utils/helper/lifecycle_event_helper.dart';
 part '../utils/helper/route_helper.dart';
+part '../utils/helper/scroll_helper.dart';
+part '../utils/helper/text_input_helper.dart';
 part '../utils/helper/toast_helper.dart';
 part '../utils/helper/tools_helper.dart';
-part '../utils/helper/text_input_helper.dart';